+++ crates/jsonschema/src/keywords/any_of.rs
@@ -3,7 +3,7 @@ use crate::{
     error::{error, no_error, ErrorIterator, ValidationError},
     node::SchemaNode,
     paths::{LazyLocation, Location},
-    primitive_type::PrimitiveType,
+    types::JsonType,
     validator::{PartialApplication, Validate},
 };
 use serde_json::{Map, Value};
@@ -35,7 +35,7 @@ impl AnyOfValidator {
                 Location::new(),
                 ctx.location().clone(),
                 schema,
-                PrimitiveType::Array,
+                JsonType::Array,
             ))
         }
     }