+++ crates/jsonschema/src/keywords/properties.rs
@@ -5,7 +5,7 @@ use crate::{
     node::SchemaNode,
     output::BasicOutput,
     paths::{LazyLocation, Location},
-    primitive_type::PrimitiveType,
+    types::JsonType,
     validator::{PartialApplication, Validate},
 };
 use serde_json::{Map, Value};
@@ -34,7 +34,7 @@ impl PropertiesValidator {
                 Location::new(),
                 ctx.location().clone(),
                 schema,
-                PrimitiveType::Object,
+                JsonType::Object,
             )),
         }
     }