+++ crates/jsonschema/src/keywords/type_.rs
@@ -3,28 +3,28 @@ use crate::{
     error::ValidationError,
     keywords::CompilationResult,
     paths::Location,
-    primitive_type::{PrimitiveType, PrimitiveTypesBitMap},
+    types::{JsonType, JsonTypeSet},
     validator::Validate,
 };
 use serde_json::{json, Map, Number, Value};
-use std::convert::TryFrom;
+use std::str::FromStr;
 
 use crate::paths::LazyLocation;
 
 pub(crate) struct MultipleTypesValidator {
-    types: PrimitiveTypesBitMap,
+    types: JsonTypeSet,
     location: Location,
 }
 
 impl MultipleTypesValidator {
     #[inline]
     pub(crate) fn compile(items: &[Value], location: Location) -> CompilationResult {
-        let mut types = PrimitiveTypesBitMap::new();
+        let mut types = JsonTypeSet::new();
         for item in items {
             match item {
                 Value::String(string) => {
-                    if let Ok(primitive_type) = PrimitiveType::try_from(string.as_str()) {
-                        types |= primitive_type;
+                    if let Ok(ty) = JsonType::from_str(string.as_str()) {
+                        types = types.insert(ty);
                     } else {
                         return Err(ValidationError::enumeration(
                             Location::new(),
@@ -41,7 +41,7 @@ impl MultipleTypesValidator {
                         location,
                         Location::new(),
                         item,
-                        PrimitiveType::String,
+                        JsonType::String,
                     ))
                 }
             }
@@ -52,17 +52,7 @@ impl MultipleTypesValidator {
 
 impl Validate for MultipleTypesValidator {
     fn is_valid(&self, instance: &Value) -> bool {
-        match instance {
-            Value::Array(_) => self.types.contains_type(PrimitiveType::Array),
-            Value::Bool(_) => self.types.contains_type(PrimitiveType::Boolean),
-            Value::Null => self.types.contains_type(PrimitiveType::Null),
-            Value::Number(num) => {
-                self.types.contains_type(PrimitiveType::Number)
-                    || (self.types.contains_type(PrimitiveType::Integer) && is_integer(num))
-            }
-            Value::Object(_) => self.types.contains_type(PrimitiveType::Object),
-            Value::String(_) => self.types.contains_type(PrimitiveType::String),
-        }
+        self.types.contains_value_type(instance)
     }
     fn validate<'i>(
         &self,
@@ -109,7 +99,7 @@ impl Validate for NullTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::Null,
+                JsonType::Null,
             ))
         }
     }
@@ -142,7 +132,7 @@ impl Validate for BooleanTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::Boolean,
+                JsonType::Boolean,
             ))
         }
     }
@@ -176,7 +166,7 @@ impl Validate for StringTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::String,
+                JsonType::String,
             ))
         }
     }
@@ -210,7 +200,7 @@ impl Validate for ArrayTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::Array,
+                JsonType::Array,
             ))
         }
     }
@@ -243,7 +233,7 @@ impl Validate for ObjectTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::Object,
+                JsonType::Object,
             ))
         }
     }
@@ -276,7 +266,7 @@ impl Validate for NumberTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::Number,
+                JsonType::Number,
             ))
         }
     }
@@ -313,7 +303,7 @@ impl Validate for IntegerTypeValidator {
                 self.location.clone(),
                 location.into(),
                 instance,
-                PrimitiveType::Integer,
+                JsonType::Integer,
             ))
         }
     }
@@ -342,7 +332,7 @@ pub(crate) fn compile<'a>(
                         Location::new(),
                         location,
                         item,
-                        PrimitiveType::String,
+                        JsonType::String,
                     )))
                 }
             } else {
@@ -353,9 +343,9 @@ pub(crate) fn compile<'a>(
             Location::new(),
             ctx.location().clone(),
             schema,
-            PrimitiveTypesBitMap::new()
-                .add_type(PrimitiveType::String)
-                .add_type(PrimitiveType::Array),
+            JsonTypeSet::new()
+                .insert(JsonType::String)
+                .insert(JsonType::Array),
         ))),
     }
 }
@@ -365,14 +355,14 @@ fn compile_single_type<'a>(
     location: Location,
     instance: &'a Value,
 ) -> CompilationResult<'a> {
-    match PrimitiveType::try_from(item) {
-        Ok(PrimitiveType::Array) => ArrayTypeValidator::compile(location),
-        Ok(PrimitiveType::Boolean) => BooleanTypeValidator::compile(location),
-        Ok(PrimitiveType::Integer) => IntegerTypeValidator::compile(location),
-        Ok(PrimitiveType::Null) => NullTypeValidator::compile(location),
-        Ok(PrimitiveType::Number) => NumberTypeValidator::compile(location),
-        Ok(PrimitiveType::Object) => ObjectTypeValidator::compile(location),
-        Ok(PrimitiveType::String) => StringTypeValidator::compile(location),
+    match JsonType::from_str(item) {
+        Ok(JsonType::Array) => ArrayTypeValidator::compile(location),
+        Ok(JsonType::Boolean) => BooleanTypeValidator::compile(location),
+        Ok(JsonType::Integer) => IntegerTypeValidator::compile(location),
+        Ok(JsonType::Null) => NullTypeValidator::compile(location),
+        Ok(JsonType::Number) => NumberTypeValidator::compile(location),
+        Ok(JsonType::Object) => ObjectTypeValidator::compile(location),
+        Ok(JsonType::String) => StringTypeValidator::compile(location),
         Err(()) => Err(ValidationError::custom(
             Location::new(),
             location,